#320 rollback `vault_identity_group_policy`, default to exclusive
